reducing code usage
